+++ tests/test_metrics.py
@@ -223,3 +223,31 @@ class MetricsTests(TestCase):
             {'is_leaf': True, 'path': 'collectd.load'},
             {'is_leaf': True, 'path': 'collectd.memory'},
         ]})
+
+    def test_metrics_index(self):
+        url = '/metrics/index.json'
+        response = self.app.get(url)
+        self.assertJSON(response, [])
+        self.assertEqual(response.headers['Content-Type'], 'application/json')
+
+        response = self.app.get(url, query_string={'jsonp': 'foo'})
+        self.assertEqual(response.data, b'foo([])')
+        self.assertEqual(response.headers['Content-Type'], 'text/javascript')
+
+        parent = os.path.join(WHISPER_DIR, 'collectd')
+        os.makedirs(parent)
+
+        for metric in ['load', 'memory', 'cpu']:
+            db = os.path.join(parent, '{0}.wsp'.format(metric))
+            whisper.create(db, [(1, 60)])
+
+        response = self.app.get(url)
+        self.assertJSON(response, [
+            u'collectd.cpu',
+            u'collectd.load',
+            u'collectd.memory',
+        ])
+        response = self.app.get(url, query_string={'jsonp': 'bar'})
+        self.assertEqual(
+            response.data,
+            b'bar(["collectd.cpu", "collectd.load", "collectd.memory"])')